pipeline/requirements.txt
# Core
boto3==1.35.76
botocore==1.35.76

# PDF parsing
unstructured[pdf]==0.16.10
unstructured-inference==0.8.1    # Only needed for hi_res strategy
pikepdf==9.3.0
pdfminer.six==20231228

# Table parsing
beautifulsoup4==4.12.3
lxml==5.3.0

# Token counting
tiktoken==0.8.0

# Embeddings + vector DB
openai>=1.30.0
psycopg2-binary>=2.9.9

# Utilities
python-dotenv==1.0.1
tqdm==4.67.1
